+++ CHANGELOG.rst
@@ -4,7 +4,9 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat: reserved keys will be informed to developer to avoid using them as a scene key
 * 🎉 feat: change to previous scene feature added, show_first_scene transition and reset_scene_manager functions added
+
 1.2.0 (2022-03-16)
 ------------------